# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ebby Sam Luxury API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CreateCategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  images: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  categorySlug?: string | null;
  stock: number;
  featured: boolean;
  /** @nullable */
  badge?: string | null;
  sizes: string[];
  colors: string[];
  /** @nullable */
  material?: string | null;
  /** @nullable */
  sku?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductInput {
  name: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  stock?: number;
  featured?: boolean;
  /** @nullable */
  badge?: string | null;
  sizes?: string[];
  colors?: string[];
  /** @nullable */
  material?: string | null;
  /** @nullable */
  sku?: string | null;
}

export interface UpdateProductInput {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl?: string;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  stock?: number;
  featured?: boolean;
  /** @nullable */
  badge?: string | null;
  sizes?: string[];
  colors?: string[];
  /** @nullable */
  material?: string | null;
  /** @nullable */
  sku?: string | null;
}

export interface UploadUrlInput {
  filename: string;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  publicUrl: string;
  key: string;
}

export interface AdminStats {
  totalProducts: number;
  totalCategories: number;
  featuredProducts: number;
  outOfStock: number;
}

export type GetProductsParams = {
  category?: string;
  featured?: string;
};