# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ebby Sam Luxury API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product management
  - name: categories
    description: Category management
  - name: orders
    description: Order management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: featured
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /upload-url:
    post:
      operationId: getUploadUrl
      tags: [products]
      summary: Get a presigned URL for image upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [products]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - createdAt

    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
        - slug

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categorySlug:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        badge:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        material:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - images
        - stock
        - featured
        - sizes
        - colors
        - createdAt
        - updatedAt

    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        badge:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        material:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
      required:
        - name
        - description
        - price
        - imageUrl

    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        badge:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        material:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]

    UploadUrlInput:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
      required:
        - filename
        - contentType

    UploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        publicUrl:
          type: string
        key:
          type: string
      required:
        - uploadUrl
        - publicUrl
        - key

    AdminStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        featuredProducts:
          type: integer
        outOfStock:
          type: integer
      required:
        - totalProducts
        - totalCategories
        - featuredProducts
        - outOfStock