# lib/api-zod/src/generated/types/createProductInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ebby Sam Luxury API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductInput {
  name: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  stock?: number;
  featured?: boolean;
  /** @nullable */
  badge?: string | null;
  sizes?: string[];
  colors?: string[];
  /** @nullable */
  material?: string | null;
  /** @nullable */
  sku?: string | null;
}