# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ebby Sam Luxury API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl: string;
  images: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  categorySlug?: string | null;
  stock: number;
  featured: boolean;
  /** @nullable */
  badge?: string | null;
  sizes: string[];
  colors: string[];
  /** @nullable */
  material?: string | null;
  /** @nullable */
  sku?: string | null;
  createdAt: Date;
  updatedAt: Date;
}