# lib/api-zod/src/generated/types/uploadUrlInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ebby Sam Luxury API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlInput {
  filename: string;
  contentType: string;
}